data_analytics/visualizations/requirements.txt
streamlit
pandas
python-dotenv
hvplot
holoviews
plost